[Core] moved OrderShippingStates to Core namespace
